pyproject.toml
# Copyright 2025 Canonical Ltd.
# See LICENSE file for licensing details.
[project]
name = "pyroscope-k8s"
version = "0.1"  # this is in fact irrelevant
requires-python = "~=3.12.0"


[project.optional-dependencies]
dev = [
    #INTEGRATION TESTS
    "jubilant",
    "pytest-jubilant~=1.0",
    "minio",
    "requests",
    "tenacity",
    "sh",
    "pytest-bdd",
    # for scripts/profilegen.py
    "grpcio",
    "protobuf",
    "opentelemetry-proto",
]


[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"
markers = ["setup", "teardown", "unit", "integration"]
bdd_features_base_dir = "features"


